[BUGFIX] Delete index documents without available site
Resolves: #3769
Releases: main, 11.5

+++ Classes/Domain/Index/Queue/GarbageRemover/AbstractStrategy.php
@@ -95,6 +95,10 @@ abstract class AbstractStrategy
             try {
                 $site = $indexQueueItem->getSite();
             } catch (InvalidArgumentException) {
+                $site = null;
+            }
+
+            if ($site === null) {
                 $this->queue->deleteItem($indexQueueItem->getType(), $indexQueueItem->getIndexQueueUid());
                 continue;
             }
